requirements.txt
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
etelemetry==0.3.1
filelock==3.15.4
fitz==0.0.1.dev0
httplib2==0.22.0
idna==3.7
isodate==0.6.1
looseversion==1.3.0
lxml==5.2.2
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pillow==10.3.0
prov==2.0.1
pydot==3.0.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==6.3.2
reportlab==4.2.0
requests==2.32.3
scipy==1.14.0
simplejson==3.19.2
six==1.16.0
tk==0.1.0
traits==6.3.2
tzdata==2024.1
urllib3==2.2.2
